Early Stages of working Sheep Temp changes
Aside from a bit of reorganizing, the main changes took place in EvolvingSheepEntity.java and DamageSourceExt.java. The latter is not a mixin and simply added two damage sources to the game that are relevant to temperature--Heatstroke and Hypothermia. Many changes occurred in the sheep entity extension; sheep now check the temperature around them every 100 ticks, and a rudimentary method for checking if they should be damaged by that temp is in place (currently all temps cause damage). They also have an "age", which increases every 100 ticks, prompting them to eventually reproduce (asexually) and die of old age.