Added Wool-Related methods to sheep and created WoolType to interact with the entity
Using injections into many methods contained in SheepEntity for wool color purposes, similar methods of data storage were implemented for wool type. To work with these different wool types (and hopefully to connect them to different models for the sheep), a WoolType object was created based largely on the characteristics of DyeColor, which uses the same principles for coloration. Also, notably, sections (referred to as regions) were added to EvolvingSheepEntity to break up the code into different parts, as they all have to interact with one another and therefore must remain in one mixin file.