Introduction of varying wool models
The next step definitely needs to be figuring a better way to organize all these files, so until then I apologize for the influx.
This commit connects the differences in WoolType that have already been established to complementary sheep models, which are now successfully rendered into the world. This includes the creation of a new model ID, SHEEP_THICK, which makes a chonky sheep.