Chat-related update
A new file called ChatExt can now be invoked to send a message to the in-game chat, which can be useful for debugging but also the player experience. For example, when sheep die due to causes that we have customized, a death message appears in chat (this might get annoying, if so just comment out the line indicated in the onDeath method of EvolvingSheepEntity). This also involved additions of death messages and names to the lang file.